Fix Vercel build: wrap login searchParams in Suspense
