+++ src/app/login/page.tsx
@@ -1,116 +1,17 @@
-'use client'
-
-import { useState, FormEvent } from 'react'
-import { useRouter, useSearchParams } from 'next/navigation'
-import { Input } from '@/components/Input'
-import { Button } from '@/components/Button'
-
-function isSafeNextPath(next: string): boolean {
-  if (!next.startsWith('/')) return false
-  if (next.startsWith('//')) return false
-  return true
-}
+import { Suspense } from 'react'
+import { LoginClient } from './LoginClient'
+import { Spinner } from '@/components/Spinner'
 
 export default function LoginPage() {
-  const router = useRouter()
-  const searchParams = useSearchParams()
-  const [email, setEmail] = useState('')
-  const [password, setPassword] = useState('')
-  const [loading, setLoading] = useState(false)
-  const [error, setError] = useState('')
-
-  async function handleSubmit(e: FormEvent) {
-    e.preventDefault()
-    setError('')
-    setLoading(true)
-
-    try {
-      const response = await fetch('/api/auth/login', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ email, password }),
-      })
-
-      const data = await response.json()
-
-      if (!response.ok) {
-        throw new Error(data.error || 'Login failed')
-      }
-
-      const next = searchParams.get('next')
-      if (next && isSafeNextPath(next)) {
-        router.push(next)
-        return
-      }
-
-      router.push(data.redirectUrl)
-    } catch (err: any) {
-      setError(err.message || 'An error occurred')
-      setLoading(false)
-    }
-  }
-
   return (
-    <div className="min-h-screen flex items-center justify-center p-4 bg-gray-50">
-      <div className="max-w-md w-full bg-white rounded-lg shadow-lg p-8">
-        <h1 className="text-2xl font-bold text-gray-900 mb-6">
-          Login to Pika
-        </h1>
-
-        <form onSubmit={handleSubmit}>
-          <Input
-            label="School Email"
-            type="email"
-            placeholder="number@gapps.yrdsb.ca"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-            disabled={loading}
-            className="mb-4"
-          />
-
-          <Input
-            label="Password"
-            type="password"
-            placeholder="Enter your password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-            disabled={loading}
-            error={error}
-          />
-
-          <div className="mt-2 text-right">
-            <button
-              type="button"
-              onClick={() => router.push('/forgot-password')}
-              className="text-sm text-blue-600 hover:underline"
-            >
-              Forgot password?
-            </button>
-          </div>
-
-          <Button
-            type="submit"
-            className="w-full mt-6"
-            disabled={loading || !email || !password}
-          >
-            {loading ? 'Logging in...' : 'Login'}
-          </Button>
-        </form>
-
-        <div className="mt-6 text-center">
-          <p className="text-sm text-gray-600">
-            Don&apos;t have an account?{' '}
-            <button
-              onClick={() => router.push('/signup')}
-              className="text-blue-600 hover:underline font-medium"
-            >
-              Sign up
-            </button>
-          </p>
+    <Suspense
+      fallback={
+        <div className="min-h-screen flex items-center justify-center">
+          <Spinner size="lg" />
         </div>
-      </div>
-    </div>
+      }
+    >
+      <LoginClient />
+    </Suspense>
   )
 }
